Moved files to new folder and corrected urls in intern-page
