+++ src/pages/internt/coffee_machine.png
